+++ mountpoint-s3-client/src/lib.rs
@@ -40,6 +40,7 @@
 //!
 //! [awscrt]: https://docs.aws.amazon.com/sdkref/latest/guide/common-runtime.html
 
+#![deny(clippy::undocumented_unsafe_blocks)]
 // Make async trait docs not-ugly on docs.rs (https://github.com/dtolnay/async-trait/issues/213)
 #![cfg_attr(docsrs, feature(async_fn_in_trait))]
 
@@ -62,7 +63,8 @@ pub mod error_metadata;
 pub use object_client::{ObjectClient, PutObjectRequest};
 
 pub use s3_crt_client::{
-    OnTelemetry, S3CrtClient, S3RequestError, get_object::S3GetObjectResponse, put_object::S3PutObjectRequest,
+    NewClientError, OnTelemetry, S3CrtClient, S3RequestError, get_object::S3GetObjectResponse,
+    put_object::S3PutObjectRequest,
 };
 
 /// Configuration for the S3 client
@@ -79,6 +81,8 @@ pub mod config {
     pub use mountpoint_s3_crt::io::io_library_init;
     #[doc(hidden)]
     pub use mountpoint_s3_crt::s3::s3_library_init;
+
+    pub use mountpoint_s3_crt::s3::pool::{MemoryPool, MemoryPoolFactory, MemoryPoolFactoryOptions};
 }
 
 /// Types used by all object clients