+++ mountpoint-s3-client/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased (v0.18.0)
 
+* Add support for custom memory pools. ([#1516](https://github.com/awslabs/mountpoint-s3/pull/1516))
 * Upgrade to Rust 2024. ([#1498](https://github.com/awslabs/mountpoint-s3/pull/1498))
 
 ## v0.17.1 (July 17, 2025)